+++ Tests/govuk_ios/govuk_ios_unit_tests/Mocks/Services/MockNotificationService.swift
@@ -1,0 +1,25 @@
+import Foundation
+import UIKit
+
+import Onboarding
+
+@testable import govuk_ios
+
+class MockNotificationService: NotificationServiceInterface {
+    func appDidFinishLaunching(launchOptions: [UIApplication.LaunchOptionsKey : Any]?) {
+
+    }
+
+    func requestPermissions(completion: @escaping () -> Void) {
+
+    }
+
+    var _stubbedShouldRequestPermission: Bool = true
+    var shouldRequestPermission: Bool {
+        _stubbedShouldRequestPermission
+    }
+
+    func fetchSlides(completion: @escaping (Result<[any OnboardingSlideViewModelInterface], any Error>) -> Void) {
+
+    }
+}
