+++ Tests/govuk_ios/govuk_ios_unit_tests/Specs/Coordinators/NotificationOnboardingCoordinatorTests.swift
@@ -45,22 +45,3 @@ class NotificationOnboardingCoordinatorTests {
         #expect(mockNavigationController._setViewControllers == nil)
     }
 }
-
-class MockNotificationService: NotificationServiceInterface {
-    func appDidFinishLaunching(launchOptions: [UIApplication.LaunchOptionsKey : Any]?) {
-
-    }
-    
-    func requestPermissions(completion: @escaping () -> Void) {
-
-    }
-
-    var _stubbedShouldRequestPermission: Bool = true
-    var shouldRequestPermission: Bool {
-        _stubbedShouldRequestPermission
-    }
-
-    func fetchSlides(completion: @escaping (Result<[any OnboardingSlideViewModelInterface], any Error>) -> Void) {
-
-    }
-}
